Revert "Merge branch 'wextra' into fix_naming_error_relevant_for_my_thesis"
This reverts commit 2c0b9da5730b749d604093d4a8413748b379d6c0, reversing
changes made to 5fc8e4e9e4c10d8ed4d0b7f16aa4ecae5dcf5e24.

+++ src/lib/utils/lossless_predicate_cast.cpp
@@ -23,12 +23,11 @@ std::optional<float> next_float_towards(const double value, const double towards
 
   const auto casted_value = static_cast<float>(value);
 
-  if ((static_cast<double>(casted_value) < value && towards < value) ||
-      (static_cast<double>(casted_value) > value && towards > value)) {
+  if ((casted_value < value && towards < value) || (casted_value > value && towards > value)) {
     return casted_value;
   }
 
-  const float next = std::nexttowardf(casted_value, static_cast<long double>(towards));
+  const float next = std::nexttowardf(casted_value, towards);
   // Maybe someone smarter understands all the edge cases of floats. I'd rather be on the safe side.
   if (!std::isfinite(next)) {
     return std::nullopt;
